+++ src/components/Button/index.tsx
@@ -1,7 +1,7 @@
 import cn from 'classnames';
 import { ReactNode } from 'react';
 import styles from './index.module.css';
-export default function Button({ children, color = 'dark' }: { children: ReactNode, color: 'dark' | 'light' }) {
+export default function Button({ className, children, color = 'dark' }: { className?: string, children: ReactNode, color: 'dark' | 'light' }) {
 	return (
 		<button className={cn(styles['button'],
 			{