+++ src/components/Paragraph/index.tsx
@@ -1,0 +1,8 @@
+import styles from './index.module.css';
+export default function Paragraph({ text, size }: { text: string, size: number}) {
+	const fontSize = `${size} px`;
+
+	return (
+		<p style={{ fontSize }} className={styles['main']}>{text}</p>
+	);
+}